combine sensor impls
