--- com.unity.ml-agents.extensions/Runtime/Sensors/ArticulationBodySensor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a9f0629c5c4cc40e58a1cf20f6b7d26b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
